scripts/requirements.txt
python-dateutil
feedgen
